+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add "adbqr connect" command to connect to already paired devices.
+- Add "adbqr manual" command to connect using a pairing code, instead of via QR code.
 
 ## 1.1.0